Move code fix to a separate class
