--- src/Features/CSharp/Portable/CodeFixes/HideBase/HideBaseCodeFixProvider.RemoveNewKeywordAction.cs
@@ -1,79 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using System.Diagnostics;
-using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.CSharp.Extensions;
-using Microsoft.CodeAnalysis.CSharp.LanguageServices;
-
-namespace Microsoft.CodeAnalysis.CSharp.CodeFixes.HideBase
-{
-    internal partial class HideBaseCodeFixProvider
-    {
-        private class RemoveNewModifierAction : CodeActions.CodeAction
-        {
-            private readonly Document _document;
-            private readonly SyntaxNode _node;
-
-            public override string Title => CSharpFeaturesResources.Remove_new_modifier;
-
-            public RemoveNewModifierAction(Document document, SyntaxNode node)
-            {
-                _document = document;
-                _node = node;
-            }
-
-            protected override async Task<Document> GetChangedDocumentAsync(CancellationToken cancellationToken)
-            {
-                var syntaxFacts = CSharpSyntaxFacts.Instance;
-
-                var newModifier = GetNewModifier(_node);
-                Debug.Assert(newModifier != default, $"'new' keyword was not found, but diagnostic was triggered");
-
-                var newNode = _node;
-
-                if (newModifier.HasTrailingTrivia || newModifier.HasLeadingTrivia)
-                {
-                    var newModifierTrivia = newModifier.GetAllTrivia().ToSyntaxTriviaList();
-                    var previousToken = newModifier.GetPreviousToken();
-                    var nextToken = newModifier.GetNextToken();
-
-                    var sourceText = await _document.GetTextAsync(cancellationToken).ConfigureAwait(false);
-                    var isFirstTokenOnLine = newModifier.IsFirstTokenOnLine(sourceText);
-
-                    var newTrivia = new SyntaxTriviaList();
-                    if (!isFirstTokenOnLine)
-                        newTrivia = newTrivia.AddRange(previousToken.TrailingTrivia);
-                    newTrivia = newTrivia
-                        .AddRange(newModifierTrivia)
-                        .AddRange(nextToken.LeadingTrivia)
-                        .CollapseSequentialWhitespaceTrivia();
-
-                    if (isFirstTokenOnLine)
-                    {
-                        var nextTokenWithMovedTrivia = nextToken.WithLeadingTrivia(newTrivia);
-                        newNode = newNode.ReplaceToken(nextToken, nextTokenWithMovedTrivia);
-                    }
-                    else
-                    {
-                        var previousTokenWithMovedTrivia = previousToken.WithTrailingTrivia(newTrivia);
-                        newNode = newNode.ReplaceToken(previousToken, previousTokenWithMovedTrivia);
-                    }
-                }
-
-                newNode = newNode.ReplaceToken(GetNewModifier(newNode), SyntaxFactory.Token(SyntaxKind.None));
-
-                var root = await _document.GetSyntaxRootAsync(cancellationToken).ConfigureAwait(false);
-                var newRoot = root.ReplaceNode(_node, newNode);
-
-                return _document.WithSyntaxRoot(newRoot);
-
-                SyntaxToken GetNewModifier(SyntaxNode node) =>
-                    syntaxFacts.GetModifierTokens(node).FirstOrDefault(m => m.IsKind(SyntaxKind.NewKeyword));
-            }
-        }
-    }
-}
